move files to Collector
