+++ AlphaTwirl/Collector/CombineIntoList.py
