+++ AlphaTwirl/Collector/CombineIntoPandasDataFrame.py
