+++ AlphaTwirl/Collector/WritePandasDataFrameToFile.py
@@ -1,17 +1,20 @@
 # Tai Sakuma <tai.sakuma@cern.ch>
-from .mkdir_p import mkdir_p
-from .listToAlignedText import listToAlignedText
+from ..mkdir_p import mkdir_p
 import os
 
 ##__________________________________________________________________||
-class WriteListToFile(object):
+class WritePandasDataFrameToFile(object):
     def __init__(self, outPath):
         self._outPath = outPath
 
     def deliver(self, results):
         if results is None: return
         f = self._open(self._outPath)
-        f.write(listToAlignedText(results))
+        if len(results.index) == 0:
+            f.write(" ".join([i for i in results.columns]) + "\n")
+        else:
+            results.to_string(f, index = False)
+            f.write("\n")
         self._close(f)
 
     def _open(self, path):
