+++ AlphaTwirl/Collector/__init__.py
@@ -1,0 +1,17 @@
+from Combine import Combine
+from CombineIntoList import CombineIntoList, countsToList, combinedToList
+from WriteListToFile import WriteListToFile
+
+##__________________________________________________________________||
+hasPandas = False
+try:
+    import pandas
+    hasPandas = True
+except ImportError:
+    pass
+
+if hasPandas:
+    from CombineIntoPandasDataFrame import CombineIntoPandasDataFrame
+    from WritePandasDataFrameToFile import WritePandasDataFrameToFile
+
+##__________________________________________________________________||
