+++ AlphaTwirl/__init__.py
@@ -1,5 +1,6 @@
 import Aggregate
 import Binning
+import Collector
 import Concurrently
 import Configure
 import Events
@@ -7,27 +8,12 @@ import HeppyResult
 import Loop
 import ProgressBar
 import Summary
-from Combine import Combine
-from CombineIntoList import CombineIntoList, countsToList, combinedToList
-from WriteListToFile import WriteListToFile
 import Loop
 import ProgressBar
 import Aggregate
 import Concurrently
 from mkdir_p import mkdir_p
 from listToAlignedText import listToAlignedText
-
-##__________________________________________________________________||
-hasPandas = False
-try:
-    import pandas
-    hasPandas = True
-except ImportError:
-    pass
-
-if hasPandas:
-    from CombineIntoPandasDataFrame import CombineIntoPandasDataFrame
-    from WritePandasDataFrameToFile import WritePandasDataFrameToFile
 
 ##__________________________________________________________________||
 import logging
