+++ tests/unit/test_Import_01.py
@@ -29,8 +29,8 @@ class TestImport_01(unittest.TestCase):
         self.assertTrue(inspect.isfunction(AlphaTwirl.Aggregate.stack_counts_categories))
         self.assertTrue(inspect.isfunction(AlphaTwirl.Aggregate.sumOverCategories))
 
-        self.assertTrue(inspect.isclass(AlphaTwirl.CombineIntoPandasDataFrame))
-        self.assertTrue(inspect.isclass(AlphaTwirl.WritePandasDataFrameToFile))
+        self.assertTrue(inspect.isclass(AlphaTwirl.Collector.CombineIntoPandasDataFrame))
+        self.assertTrue(inspect.isclass(AlphaTwirl.Collector.WritePandasDataFrameToFile))
 
     @unittest.skipUnless(hasROOT, "has no ROOT")
     def test_with_ROOT(self):
@@ -50,8 +50,9 @@ class TestImport_01(unittest.TestCase):
         self.assertTrue(inspect.isclass(AlphaTwirl.Binning.Round))
         self.assertTrue(inspect.isclass(AlphaTwirl.Binning.RoundLog))
         self.assertTrue(inspect.isclass(AlphaTwirl.Concurrently.CommunicationChannel))
-        self.assertTrue(inspect.isclass(AlphaTwirl.Combine))
-        self.assertTrue(inspect.isclass(AlphaTwirl.CombineIntoList))
+        self.assertTrue(inspect.isclass(AlphaTwirl.Collector.Combine))
+        self.assertTrue(inspect.isclass(AlphaTwirl.Collector.CombineIntoList))
+        self.assertTrue(inspect.isclass(AlphaTwirl.Collector.WriteListToFile))
         self.assertTrue(inspect.isclass(AlphaTwirl.Summary.Reader))
         self.assertTrue(inspect.isclass(AlphaTwirl.Summary.Count))
         self.assertTrue(inspect.isclass(AlphaTwirl.Summary.NextKeyComposer))
@@ -82,6 +83,5 @@ class TestImport_01(unittest.TestCase):
         self.assertTrue(inspect.isclass(AlphaTwirl.ProgressBar.NullProgressMonitor))
         self.assertTrue(inspect.isclass(AlphaTwirl.ProgressBar.BProgressMonitor))
         self.assertTrue(inspect.isclass(AlphaTwirl.ProgressBar.ProgressReportPickup))
-        self.assertTrue(inspect.isclass(AlphaTwirl.WriteListToFile))
 
 ##________________________________._________________________________||
